refactory : View 클래스 InputView와 OutputView로 분리
